test: extract MockPrivacyDataExporter.swift from ExportTestSupport.swift
Splits the `MockPrivacyDataExporter` test double (~30 lines) out of
ExportTestSupport.swift into its own file, following the pattern of
#743 MockArtifactsService.

Byte-identical move of lines 68-98. Both files retain their test imports.

ExportTestSupport.swift: 98 → 66 lines. Tests: 667.
